+++ Changelog.md
@@ -14,6 +14,7 @@
 - Added "Fullscreen" button in the footer, or **Alt/Ctrl**+**F** or **F11** hotkey
 - Added option "Hold Escape button to exit UI fullscreen"
 - F5 and Ctrl+R hotkeys are overridden to use footer's "Reload" button (with autosave)
+- Presets batch mode by default if too many presets (filter appear threshold is used)
 
 ### 2.2 – Improved files open and LLMs support
 - Support boolean primitive nodes (e.g. useful for "thinking" in text generation node)